hicode-api/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
pydantic==2.9.0
python-dotenv==1.0.0

# HICode dependencies
datasets==3.3.2
openai==1.63.2
sentence-transformers==3.4.1
tqdm==4.67.1
torch==2.5.1
